+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- support for TypedData for CSV export, [PR-17](https://github.com/panda-official/DriftCLI/pull/17)
+
 ## 0.8.0 - 2023-07-11
 
 ### Added